added delete and search functionality

+++ projectRepos/studentManagement/script.js
@@ -12,11 +12,26 @@ const addStudentBtnEl = document.getElementById('addStudent')
 const showStudentBtnEl = document.getElementById('showStudent')
 const textDisplayEl = document.getElementById('textDisplay')
 const recordDisplayEl = document.getElementById('recordDisplay')
+const searchInputEl = document.getElementById('searchInput')
+
+
 window.onload = () => {
     studentsArr.forEach((e => renderItems(e)))
 }
 
+searchInputEl.addEventListener('input', () => {
+    const query = searchInputEl.value.toLowerCase()
 
+    recordDisplayEl.innerHTML = ''; 
+    recordDisplayEl.style.display = 'block';
+
+
+    const filteredStudents = studentsArr.filter(s => s.name.toLowerCase().includes(query) || 
+        s.id.toString().includes(query))
+
+    filteredStudents.forEach(s => renderItems(s))
+    }
+)
 addStudentBtnEl.addEventListener('click',  addStudent)
 showStudentBtnEl.addEventListener('click' , showStudent)
 
@@ -51,7 +66,11 @@ function addStudent(){
     }
     const newStudent = new Student(userId, userName, userAge,userGrade,
         userHobby)
-        
+    const userExists = studentsArr.some(s => s.id === userId)
+    if(userExists){
+        alert("You've entered a user Id that already existed")
+        return
+    }
     studentsArr.push(newStudent)
     localStorage.setItem("students", JSON.stringify(studentsArr))
     renderItems(newStudent)
@@ -59,11 +78,24 @@ function addStudent(){
 
 function renderItems(students){
 const myDiv = document.createElement("div");
-    recordDisplayEl.appendChild(myDiv);
+const deleteBtn = document.createElement("button");
     myDiv.textContent = `ID: ${students.id} | Name: ${students.name} | Age: ${students.age} | Grade: ${students.grade} |
     Hobbies: ${students.hobbies}`
+    deleteBtn.textContent = "delete"
+
+    deleteBtn.onclick = () => {
+      studentsArr = studentsArr.filter(s => s.id !== students.id)
+       localStorage.setItem("students", JSON.stringify(studentsArr))
+        myDiv.remove()
+        deleteBtn.remove()
+    }
+    recordDisplayEl.appendChild(myDiv);
+    recordDisplayEl.appendChild(deleteBtn)
 }
+
 
 function showStudent(){
     recordDisplayEl.style.display = recordDisplayEl.style.display === 'none' ? 'block' : 'none'  
+    recordDisplayEl.innerHTML = ''; 
+    studentsArr.forEach(s => renderItems(s))
 }
